:recycle: refactor(flux): 액션 상수 위치를 js에서 flux로 변경
