+++ src/js/flux/actions.js
@@ -1,0 +1,7 @@
+const ACTION = {
+  PURCHASE_LOTTO: 'purchase-lotto',
+  TOGGLE_LOTTO_LIST: 'toggle-lotto-list',
+  SET_WINNING_NUMBERS: 'set-winning-numbers',
+};
+
+export default ACTION;
