+++ src/js/flux/reducer.js
@@ -1,4 +1,5 @@
-import { ACTION, LOTTO } from '../constants';
+import ACTION from './actions';
+import { LOTTO } from '../constants';
 import { pickUniqueNumbersInRange } from '../utils';
 
 const generateLottoList = (money) => {
